requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.5.3
boto3==1.4.4
botocore==1.5.95
decorator==4.1.2
docutils==0.14
dumptruck==0.1.6
enum34==1.1.6
futures==3.1.1
ipdb==0.10.3
ipython==5.4.1
ipython-genutils==0.2.0
jmespath==0.9.3
nose==1.3.7
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.13.0
s3transfer==0.1.10
scandir==1.5
-e git+http://github.com/openaustralia/scraperwiki-python.git@732dda1982a3b2073f6341a6a24f9df1bda77fa0#egg=scraperwiki
simplegeneric==0.8.1
six==1.10.0
slacker==0.9.42
slacker-log-handler==1.6.1
traitlets==4.3.2
wcwidth==0.1.7
